requirements.txt
customtkinter~=5.2.2
pillow~=10.3.0
bcrypt~=4.1.3
tkcalendar~=1.6.1
mysql~=0.0.3
mysql-connector-python~=8.4.0
ctktable~=1.1